Add 128bit trace_id support (#230)
* Add 128bit trace_id support

- Creates a new property in Config 'generate_128bit_trace_id' if set
  to True then 128bit trace_ids are generated
  - Adds support for environment variable JAEGER_TRACEID_128BIT. If
  the value is 'true' then 128bit trace_ids are generated
- Tracer takes variable generate_128bit_trace_id, if the value is True
  then max_trace_id_bits is set to constants._max_trace_id_bits
  - Adds new function _random_id(bitsize)
  - Deprecate random_id() in favor of _random_id(bitsize)
- Deprecate MAX_ID_BITS in favor of _max_trace_id_bits and _max_id_bits
- Set low and high bits of traceId correctly
  - Uses _id_to_high() and _id_to_low() to do that


* Add tests for 128bit trace_id support


* Update ProbabilisticSampler to trim the trace_id to 64bits


* Update test_round_trip to test with 128bit trace_id

- Uses itertools.product to create a list with all possible
  combinations for tracer i.e. tracer with 64bit trace_id and tracer
  with 128bit trace_id


* Add test_inject_with_128bit_trace_id to test_codecs.py

- This tests the behavior of inject for different codes when it
  receives 64bit/128bit trace_id on wire when tracer is configured to
  genereate 64/128bit trace_ids
- Also tests the behavior of inject when tracer is configured to
  generate 64/128bit trace_ids

+++ tests/test_sampler.py
@@ -195,7 +195,7 @@ def test_guaranteed_throughput_probabilistic_sampler():
     sampled, tags = sampler.is_sampled(MAX_INT - 10)
     assert sampled
     assert tags == get_tags('probabilistic', 0.51)
-    sampled, tags = sampler.is_sampled(MAX_INT + (MAX_INT / 4))
+    sampled, tags = sampler.is_sampled(int(MAX_INT + (MAX_INT / 4)))
     assert sampled
     assert tags == get_tags('lowerbound', 0.51)
 
@@ -234,7 +234,7 @@ def test_adaptive_sampler():
         # Move time forward by a second to guarantee the rate limiter has enough credits
         mock_time.side_effect = lambda: ts + 1
 
-        sampled, tags = sampler.is_sampled(MAX_INT + (MAX_INT / 4), 'new_op')
+        sampled, tags = sampler.is_sampled(int(MAX_INT + (MAX_INT / 4)), 'new_op')
         assert sampled
         assert tags == get_tags('lowerbound', 0.51)
 
@@ -243,7 +243,7 @@ def test_adaptive_sampler():
     sampled, tags = sampler.is_sampled(MAX_INT - 10, 'new_op_2')
     assert sampled
     assert tags == get_tags('probabilistic', 0.51)
-    sampled, _ = sampler.is_sampled(MAX_INT + (MAX_INT / 4), 'new_op_2')
+    sampled, _ = sampler.is_sampled(int(MAX_INT + (MAX_INT / 4)), 'new_op_2')
     assert not sampled
     assert '%s' % sampler == 'AdaptiveSampler(0.510000, 3.000000, 2)'
